Move daq::mutex to coretypes
